+++ src/mnextend/__init__.py
@@ -2,8 +2,17 @@
 #
 # License: BSD (3-clause)
 
+from mnextend.iclabel import IC_LABELS, run_iclabel
 from mnextend.io.readers import read_epochs, read_raw
 from mnextend.io.utils import split_name_ext
 from mnextend.io.writers import write_epochs, write_raw
 
-__all__ = ["read_epochs", "read_raw", "split_name_ext", "write_epochs", "write_raw"]
+__all__ = [
+    "IC_LABELS",
+    "read_epochs",
+    "read_raw",
+    "run_iclabel",
+    "split_name_ext",
+    "write_epochs",
+    "write_raw",
+]